refactor: 도메인 및 뷰 패키지 이동
