+++ src/main/java/racingcar/domain/RacingCars.java
@@ -1,7 +1,5 @@
 package racingcar.domain;
 
-import racingcar.RacingCar;
-
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
