+++ src/test/java/racingcar/domain/RacingCarTest.java
@@ -1,11 +1,10 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-
-import java.util.Arrays;
+import racingcar.domain.RacingCar;
 
 import static org.assertj.core.api.Assertions.*;
